+++ app/javascript/packs/components/loader.jsx
@@ -1,0 +1,115 @@
+import LinearProgress from "@material-ui/core/LinearProgress";
+import Fade from "@material-ui/core/Fade";
+import InvoiceList from "./invoice_list";
+import React from "react";
+import Snackbar from '@material-ui/core/Snackbar';
+
+// This contains the kick-off point for an asynchronous simulation:
+//
+//       app/views/setup/invoices/async_invoice.html.erb
+//
+// Will include this file, it will also render a root element with the
+// URL to poll in it as a data-element.
+//
+// On load this page will poll that URL every 1 second, and once we
+// have a payload URL to a S3 blob, we will fetch that and create the
+// various invoices.
+const getJSON = async function(url) {
+  const response = await fetch(url);
+  const json = await response.json();
+  return json;
+};
+
+class Loader extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+
+  state = {
+    loading: true,
+    error: null,
+    jsonPayload: { invoices: [] },
+  };
+
+  async componentDidMount() {
+    const url = this.props.url;
+    this.timer = setInterval(() => this.apiFetch(url), 1000);
+  }
+
+  componentWillUnmount() {
+    clearInterval(this.timer);
+    this.timer = null;
+  }
+
+  async fetchJSON(url) {
+    try {
+      const response = await fetch(url, {});
+      const json = await response.json();
+      return {success: true, payload: json};
+    } catch (error) {
+      return {success: false, payload: error};
+    }
+  }
+
+  async apiFetch(url) {
+    const response = await this.fetchJSON(url);
+    if (response.success) {
+      if (!response.payload.data.attributes.isAlive) {
+        clearInterval(this.timer);
+        this.fetchSimulationResult(response.payload.data.attributes.resultUrl);
+      } else {
+        console.log(response.payload.data);
+        // I'm still alive
+      }
+    } else {
+      clearInterval(this.timer);
+      this.showError(response.payload);
+    }
+  }
+
+  async fetchSimulationResult(url) {
+    const response = await this.fetchJSON(url);
+    if (response.success) {
+      this.setState({
+        loading: false,
+        jsonPayload: response.payload,
+      });
+    } else {
+      this.showError(response.payload);
+    }
+  }
+
+  showError(error) {
+    this.setState({
+      loading: false,
+      errorMessage: error.message,
+    });
+  }
+
+  render() {
+    const { loading, errorMessage, jsonPayload } = this.state;
+    const isLoaded = !loading;
+    const hasError = !!errorMessage;
+    const isSuccess = isLoaded && !hasError;
+
+    return (
+      <>
+        <Fade
+          in={loading}
+          unmountOnExit
+        >
+          <LinearProgress variant="query" />
+        </Fade>
+        <Snackbar
+          anchorOrigin={{vertical: 'top', horizontal: 'right' }}
+          open={isLoaded && hasError}
+          autoHideDuration={6000}
+          message={<span id="message-id">Error: {errorMessage}</span>}
+        />
+        {isSuccess && <InvoiceList key="list" invoices={jsonPayload.invoices} />}
+      </>
+    );
+  }
+}
+
+export default Loader;
